add tensor storage class
